requirements.txt
insightface==0.7.3
numpy==2.4.1
PySide6==6.10.1
Requests==2.32.5
ultralytics==8.4.5
fastapi==0.128.0
pydantic==2.12.5
python-dotenv==1.2.1
schedule==1.2.2
sentence_transformers==5.2.0
uvicorn==0.40.0
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
sqlalchemy==2.0.23
sqlalchemy_utils==0.41.0